Section11

+++ resources/item.py
@@ -1,21 +1,20 @@
-from sqlite3.dbapi2 import Cursor
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt, get_jwt_identity
 
 from models.item import ItemModel
 
 class Item(Resource):
     parser = reqparse.RequestParser()
     parser.add_argument("price",
-        type=float,
-        required=True,
-        help="This field cannot be left blank!"
-    )
+                        type=float,
+                        required=True,
+                        help="This field cannot be left blank!"
+                        )
     parser.add_argument("store_id",
-        type=int,
-        required=True,
-        help="Every item needs a store_id!"
-    )
+                        type=int,
+                        required=True,
+                        help="Every item needs a store_id!"
+                        )
 
     @jwt_required()
     def get(self, name):
@@ -24,6 +23,7 @@ class Item(Resource):
             return item.json()
         return {"message": "Item not found"}, 404
 
+    @jwt_required(fresh=True)
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {"message": f"An item with name {name} already exists"}, 400
@@ -38,23 +38,30 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims["is_admin"]:
+            return {"message": "Admin privilege required"}
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
-
-        return {"message": "Item deleted"}
+            return {"message": "Item deleted"}
+        else:
+            return {"message": "Item not found."}, 404
+        
     
     def put(self, name):
         data = Item.parser.parse_args()
         
         item = ItemModel.find_by_name(name)
 
-        if item is None:
-            item = ItemModel(name, **data)
-        else:
+        
+        if item:
             item.price = data["price"]
             item.store_id = data["store_id"]
+        else:
+            item = ItemModel(name, **data)
 
         item.save_to_db()
 
@@ -62,5 +69,13 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_required(optional=True)
     def get(self):
-        return {"items": [item.json() for item in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {"items": items}, 200
+        return {
+            "items": [item["name"] for item in items],
+            "message": "More data available if you log in."
+        }, 200
